refactor: rename `Events/` components
